Private headers for connection
