+++ src/router_node.c
@@ -23,6 +23,7 @@
 #include <stdbool.h>
 #include <stdlib.h>
 #include <qpid/dispatch.h>
+#include "connection_private.h"
 #include "dispatch_private.h"
 #include "entity_cache.h"
 #include "router_private.h"
@@ -70,7 +71,7 @@ static void qdr_node_connect_deliveries(qd_link_t *link, qdr_delivery_t *qdlv, p
     pn_delivery_set_context(pdlv, ref);
     qdr_delivery_set_context(qdlv, pdlv);
     qdr_delivery_incref(qdlv, "referenced by a pn_delivery");
-    
+
 }
 
 
@@ -292,7 +293,7 @@ static void log_link_message(qd_connection_t *conn, pn_link_t *pn_link, qd_messa
     if (!cf) return;
     char buf[qd_message_repr_len()];
     const char *msg_str = qd_message_oversize(msg) ? "oversize message" :
-                          qd_message_aborted(msg) ? "aborted message" : 
+                          qd_message_aborted(msg) ? "aborted message" :
                           qd_message_repr(msg, buf, sizeof(buf), cf->log_bits);
     if (msg_str) {
         const char *src = pn_terminus_get_address(pn_link_source(pn_link));
@@ -1500,7 +1501,7 @@ static void CORE_connection_activate(void *context, qdr_connection_t *conn)
 
 static void CORE_link_first_attach(void             *context,
                                    qdr_connection_t *conn,
-                                   qdr_link_t       *link, 
+                                   qdr_link_t       *link,
                                    qdr_terminus_t   *source,
                                    qdr_terminus_t   *target,
                                    qd_session_class_t ssn_class)
@@ -2056,7 +2057,7 @@ void qd_connection_log_policy_denial(qd_link_t *link, const char *text)
         if (rlink->conn) {
             c_id = rlink->conn->identity;
         }
-    }    
+    }
     qd_log(qd_policy_log_source(), QD_LOG_WARNING, "[C%"PRIu64"][L%"PRIu64"] %s",
            c_id, l_id, text);
 }
