+++ src/server_event_handlers.c
@@ -28,6 +28,7 @@
 #include <stdio.h>
 
 #include "remote_sasl.h"
+#include "connection_private.h"
 #include "server_private.h"
 
 static void handle_listener_open(qd_server_t* server, qd_listener_t* listener) {
